client/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mlx-container"
version = "0.1.0"
description = "Python client for MLX Container Toolkit — GPU-accelerated inference in Linux containers on Apple Silicon"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "AIFLOW LABS", email = "ilessio@aiflowlabs.io" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.3.0",
]

[tool.hatch.build.targets.wheel]
packages = ["mlx_container"]
